backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.26.0
langchain==0.1.9
langchain-community==0.0.38 